+++ lib/punchblock/translator/asterisk/component/input.rb
@@ -13,7 +13,7 @@ module Punchblock
           end
 
           def execute
-            @call.answer_if_not_answered
+            @call.send_progress
             initial_timeout = @component_node.initial_timeout || -1
             @inter_digit_timeout = @component_node.inter_digit_timeout || -1